TV-DPCM in JAGS
Alternative versions of the penalized TV-DPCM to use in JAGS. The model
tv_dpcm.jags is exactly equal to the model in tv_dpcm_int_5.1.stan,
which uses a random walk prior to penalize the smooth function. On the
other hand, the model tv_dpcm_jagam.jags uses the penalization from the
mcgv package. This penalization is achieved via a multivariate normal
prior on the coefficiens of the smooth function.