Revision of the simulation.
Basedon the suggestion of one reviewer, we rerun a part of the
simulation in order to save the BULK and TAIL ESS of the parameters.
This simulation was run with the conditions without missing values and
with five replications per conditions. Here, we upload the output files
of this simulation.